refactor: Move ruff_cli::resolve to own module
